update belief code and clean some old code
