+++ stormvogel/teaching/belief.py
@@ -1,11 +1,12 @@
-"""Teaching module: POMDP belief computations.
+"""Teaching module: POMDP belief type and belief computations.
 
-Provides exact Bayesian belief tracking for POMDPs with deterministic state
-observations.  All arithmetic uses :class:`~fractions.Fraction`.
+Provides the canonical :class:`Belief` type and exact Bayesian belief
+tracking for POMDPs with deterministic state observations.  All arithmetic
+uses :class:`~fractions.Fraction`.
 
 Typical usage::
 
-    from stormvogel.teaching.pomdp import initial_belief, belief_trace
+    from stormvogel.teaching.belief import Belief, initial_belief, belief_trace
 
     b0 = initial_belief(pomdp, "z")
     beliefs = belief_trace(pomdp, b0, [("b", "z"), ("a", "z_target")])
@@ -13,16 +14,68 @@ Typical usage::
 
 from __future__ import annotations
 
+from collections import defaultdict
+from collections.abc import Iterator, Mapping
 from fractions import Fraction
 from typing import TYPE_CHECKING
+from uuid import UUID
 
 if TYPE_CHECKING:
     from stormvogel.model.model import Model
     from stormvogel.model.state import State
 
-#: A belief is an exact rational probability distribution over states.
-#: Absent states have implicit probability zero.
-Belief = dict["State", Fraction]
+
+class Belief(Mapping["State", Fraction]):
+    """Exact probability distribution over POMDP states.
+
+    Implements the :class:`~collections.abc.Mapping` interface over
+    ``State → Fraction``, so ``b[s]``, ``b.get(s, 0)``, ``b.items()``,
+    etc. all work directly.  Zero-probability states are silently dropped.
+
+    :param dist: Mapping from POMDP states to their belief probabilities.
+    """
+
+    def __init__(self, dist: "dict[State, Fraction]") -> None:
+        self.dist: dict["State", Fraction] = {s: p for s, p in dist.items() if p > 0}
+        self._key: tuple[tuple[UUID, Fraction], ...] = tuple(
+            sorted(((s.state_id, p) for s, p in self.dist.items()), key=lambda x: x[0])
+        )
+
+    # --- Mapping interface ---------------------------------------------------
+
+    def __getitem__(self, key: "State") -> Fraction:
+        return self.dist[key]
+
+    def __iter__(self) -> Iterator["State"]:
+        return iter(self.dist)
+
+    def __len__(self) -> int:
+        return len(self.dist)
+
+    # --- Identity ------------------------------------------------------------
+
+    def __hash__(self) -> int:
+        return hash(self._key)
+
+    def __eq__(self, other: object) -> bool:
+        if isinstance(other, Belief):
+            return self._key == other._key
+        return NotImplemented
+
+    def __repr__(self) -> str:
+        return f"Belief({self.dist!r})"
+
+    @classmethod
+    def normalize(cls, unnorm: "dict[State, Fraction]") -> "Belief":
+        """Normalize *unnorm* to a probability distribution and return a Belief.
+
+        :param unnorm: Unnormalized weights (non-negative, at least one > 0).
+        :raises ValueError: If all weights are zero.
+        """
+        total = sum(unnorm.values(), Fraction(0))
+        if total == 0:
+            raise ValueError("Cannot normalize a zero-weight distribution.")
+        return cls({s: v / total for s, v in unnorm.items()})
 
 
 def initial_belief(pomdp: "Model", obs_alias: str) -> Belief:
@@ -47,21 +100,20 @@ def initial_belief(pomdp: "Model", obs_alias: str) -> Belief:
         pomdp.get_observation(obs_alias)
     ]
 
-    unnorm: dict["State", Fraction] = {}
+    unnorm: defaultdict["State", Fraction] = defaultdict(Fraction)
     for action, branch in pomdp.transitions[init]:
         if action is not EmptyAction:
             continue
         for prob, tgt in branch:
             if tgt in obs_states:
-                unnorm[tgt] = unnorm.get(tgt, Fraction(0)) + Fraction(prob)
+                unnorm[tgt] += Fraction(prob)
 
-    total = sum(unnorm.values(), Fraction(0))
-    if total == 0:
+    if not unnorm:
         raise ValueError(
             f"No states with observation '{obs_alias}' are reachable from the "
             f"initial state under the EmptyAction transition."
         )
-    return {s: v / total for s, v in unnorm.items()}
+    return Belief.normalize(unnorm)
 
 
 def belief_update(
@@ -92,7 +144,7 @@ def belief_update(
         pomdp.get_observation(obs_alias)
     ]
 
-    unnorm: dict["State", Fraction] = {}
+    unnorm: defaultdict["State", Fraction] = defaultdict(Fraction)
     for state, choices in pomdp.transitions.items():
         b_s = belief.get(state, Fraction(0))
         if b_s == 0:
@@ -103,15 +155,14 @@ def belief_update(
                 continue
             for prob, tgt in branch:
                 if tgt in obs_states:
-                    unnorm[tgt] = unnorm.get(tgt, Fraction(0)) + Fraction(prob) * b_s
+                    unnorm[tgt] += Fraction(prob) * b_s
 
-    total = sum(unnorm.values(), Fraction(0))
-    if total == 0:
+    if not unnorm:
         raise ValueError(
             f"Belief update failed: observation '{obs_alias}' is unreachable "
             f"from the current belief under action '{action_label}'."
         )
-    return {s: v / total for s, v in unnorm.items()}
+    return Belief.normalize(unnorm)
 
 
 def belief_trace(
